requirements.txt
tekore==4.4.0
python-vlc==3.0.16120
Pillow
